pyproject.toml
[tool.poetry]
name = "imagedb"
version = "0.1.4dev1"
description = "Store images, especially from Clipboard, in database"
authors = ["patarapolw <patarapolw@gmail.com>"]

[tool.poetry.dependencies]
python = "*"
pillow = "^5.2"
sqlalchemy = "^1.2"
nonrepeat = "^0.1.1"
flask = "^1.0"
python-slugify = "^1.2"
send2trash = "^1.5"
imagehash = "^4.0"
SQLAlchemy-Continuum = "^1.3"

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
notebook = "^5.6"
tinydb = "^3.11"
alembic = "^1.0"
